fuse grid module files into a single file
