--- honeycomb-core/src/cmap/builder/grid/mod.rs
@@ -1,10 +1,0 @@
-//! Utility for sample map generation
-//!
-//! This module contains code used to generate maps that represent grids. These have a variety
-//! of usages, most notably in tests, benchmarks, and specific algorithms.
-
-pub mod building_routines;
-pub mod descriptor;
-
-#[cfg(test)]
-mod tests;
